fix(launcher): :bug: Restore Launcher
