+++ alvr/launcher/Cargo.toml
@@ -11,9 +11,11 @@ alvr_gui_common.workspace = true
 
 anyhow = "1"
 eframe = "0.26"
+flate2 = "1"
 futures-util = "0.3.28"
 open = "5"
-reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "stream", "json"] }
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "stream", "json", "http2"] }
 serde_json = "1"
+tar = "0.4"
 tokio = { version = "1", features = ["rt-multi-thread"] }
 zip = "0.6"
